Update App page to use App.css
- Fixes navigation by making entire button area click-able
- Removes unnecessary logic to control active link by using proper routing components
- Removes entire Header component
- Fixes display of other pages (text is not clipped by header)
- Fixes whitespace above header due to percentage based margin
- Fixes inconsistent whitespace from left vs right side of screen
- Fixes issue with link wrapping clipping on narrow screens
- Fixes issue with user name text spilling outside of container element
- Remove old otherPages/ components no longer used.
