--- src/components/otherPages/About.tsx
@@ -1,14 +1,0 @@
-import * as React from 'react';
-
-export default class About extends React.Component<any, any>{
-    constructor(p: any){
-        super(p)
-    }
-    render(){
-        return (
-            <div className='dummyContent'>
-                <span className="ms-font-su dummyText">ABOUT</span>
-            </div>
-        )
-    }
-}
