--- src/components/otherPages/Docs.tsx
@@ -1,14 +1,0 @@
-import * as React from 'react';
-
-export default class Docs extends React.Component<any, any> {
-    constructor(p: any) {
-        super(p)
-    }
-    render() {
-        return (
-            <div className='dummyContent'>
-                <span className="ms-font-su dummyText">DOCS</span>
-            </div>
-        )
-    }
-}
