+++ src/containers/AppAdmin.tsx
@@ -24,7 +24,7 @@ class AppAdmin extends React.Component<Props, ComponentState> {
         display: 'Dash',
         selectedKey: 'Dash'
     }
-    
+
     renderChosenNavLink() {
         switch (this.state.display) {
             case "Settings":
@@ -65,38 +65,36 @@ class AppAdmin extends React.Component<Props, ComponentState> {
     }
     renderWithoutEmulator() {
         return (
-            <div className="container">
-                <div className="content">
-                    <div className='trainingGroundNavigationArea'>
-                        <span className="ms-font-xxl">{this.props.app.appName}</span>
-                        <div className="tgSettingsDiv">
-                            <a onClick={() => this.setArenaDisplay('Settings')}><span className="ms-Icon ms-Icon--Settings" aria-hidden="true"></span>&nbsp;&nbsp;</a>
-                            <span className="ms-font-m-plus"><a onClick={() => this.setArenaDisplay('Settings')}>Settings</a></span>
-                        </div>
-                        <div className="tgNavDiv">
-                            <Nav
-                                className="ms-font-m-plus trainingGroundNav"
-                                initialSelectedKey="Dash"
-                                selectedKey={this.state.selectedKey}
-                                groups={[{
-                                    links: [
-                                        { name: 'Dashboard', key: 'Dash', url: null, onClick: () => this.setArenaDisplay('Dash') },
-                                        { name: 'Entities', key: 'Entities', url: null, onClick: () => this.setArenaDisplay('Entities') },
-                                        { name: 'Actions', key: 'Actions', url: null, onClick: () => this.setArenaDisplay('Actions') },
-                                        { name: 'Train Dialogs', key: 'TrainDialogs', url: null, onClick: () => this.setArenaDisplay('TrainDialogs') },
-                                        { name: 'Log Dialogs', key: 'LogDialogs', url: null, onClick: () => this.setArenaDisplay('LogDialogs') }
-                                    ]
-                                }]}
-                            />
-                        </div>
-                        <div className="tgbackToAppsDiv">
-                            <Link className="backToApps" onClick={() => this.props.setDisplayMode(DisplayMode.AppList)}><span className="ms-Icon ms-Icon--Back backToApps backToAppsIcon" aria-hidden="true"></span>&nbsp;&nbsp;App List</Link>
-                        </div>
+            <div>
+                <div className='trainingGroundNavigationArea'>
+                    <span className="ms-font-xxl">{this.props.app.appName}</span>
+                    <div className="tgSettingsDiv">
+                        <a onClick={() => this.setArenaDisplay('Settings')}><span className="ms-Icon ms-Icon--Settings" aria-hidden="true"></span>&nbsp;&nbsp;</a>
+                        <span className="ms-font-m-plus"><a onClick={() => this.setArenaDisplay('Settings')}>Settings</a></span>
                     </div>
+                    <div className="tgNavDiv">
+                        <Nav
+                            className="ms-font-m-plus trainingGroundNav"
+                            initialSelectedKey="Dash"
+                            selectedKey={this.state.selectedKey}
+                            groups={[{
+                                links: [
+                                    { name: 'Dashboard', key: 'Dash', url: null, onClick: () => this.setArenaDisplay('Dash') },
+                                    { name: 'Entities', key: 'Entities', url: null, onClick: () => this.setArenaDisplay('Entities') },
+                                    { name: 'Actions', key: 'Actions', url: null, onClick: () => this.setArenaDisplay('Actions') },
+                                    { name: 'Train Dialogs', key: 'TrainDialogs', url: null, onClick: () => this.setArenaDisplay('TrainDialogs') },
+                                    { name: 'Log Dialogs', key: 'LogDialogs', url: null, onClick: () => this.setArenaDisplay('LogDialogs') }
+                                ]
+                            }]}
+                        />
+                    </div>
+                    <div className="tgbackToAppsDiv">
+                        <Link className="backToApps" onClick={() => this.props.setDisplayMode(DisplayMode.AppList)}><span className="ms-Icon ms-Icon--Back backToApps backToAppsIcon" aria-hidden="true"></span>&nbsp;&nbsp;App List</Link>
+                    </div>
+                </div>
 
-                    <div className='trainingGroundArena'>
-                        {this.renderChosenNavLink()}
-                    </div>
+                <div className='trainingGroundArena'>
+                    {this.renderChosenNavLink()}
                 </div>
             </div>
         )
@@ -104,34 +102,30 @@ class AppAdmin extends React.Component<Props, ComponentState> {
     }
     renderWithEmulator() {
         return (
-            <div className="container">
-                <div className="emulatorContent">
-                    <div className='trainingGroundNavigationArea'>
-                        <span className="ms-font-xxl">{this.props.app.appName}</span>
-                        <div className="tgSettingsDiv">
-                            <a onClick={() => this.setArenaDisplay('Settings')}><span className="ms-Icon ms-Icon--Settings" aria-hidden="true"></span>&nbsp;&nbsp;</a>
-                            <span className="ms-font-m-plus"><a onClick={() => this.setArenaDisplay('Settings')}>Settings</a></span>
-                        </div>
-                        <div className="tgNavDiv">
-                            <Nav
-                                className="ms-font-m-plus trainingGroundNav"
-                                initialSelectedKey="Dash"
-                                selectedKey={this.state.selectedKey}
-                                groups={[{
-                                    links: [
-                                        { name: 'Dashboard', key: 'Dash', url: null, onClick: () => this.setArenaDisplay('Dash') },
-                                        { name: 'Entities', key: 'Entities', url: null, onClick: () => this.setArenaDisplay('Entities') },
-                                        { name: 'Actions', key: 'Actions', url: null, onClick: () => this.setArenaDisplay('Actions') },
-                                        { name: 'Train Dialogs', key: 'TrainDialogs', url: null, onClick: () => this.setArenaDisplay('TrainDialogs') },
-                                        { name: 'Log Dialogs', key: 'LogDialogs', url: null, onClick: () => this.setArenaDisplay('LogDialogs') }
-                                    ]
-                                }]}
-                            />
-                        </div>
-                        <div className="tgbackToAppsDiv">
-                            <Link className="backToApps" onClick={() => this.props.setDisplayMode(DisplayMode.AppList)}><span className="ms-Icon ms-Icon--Back backToApps backToAppsIcon" aria-hidden="true"></span>&nbsp;&nbsp;App List</Link>
-                        </div>
-                    </div>
+            <div>
+                <span className="ms-font-xxl">{this.props.app.appName}</span>
+                <div className="tgSettingsDiv">
+                    <a onClick={() => this.setArenaDisplay('Settings')}><span className="ms-Icon ms-Icon--Settings" aria-hidden="true"></span>&nbsp;&nbsp;</a>
+                    <span className="ms-font-m-plus"><a onClick={() => this.setArenaDisplay('Settings')}>Settings</a></span>
+                </div>
+                <div className="tgNavDiv">
+                    <Nav
+                        className="ms-font-m-plus trainingGroundNav"
+                        initialSelectedKey="Dash"
+                        selectedKey={this.state.selectedKey}
+                        groups={[{
+                            links: [
+                                { name: 'Dashboard', key: 'Dash', url: null, onClick: () => this.setArenaDisplay('Dash') },
+                                { name: 'Entities', key: 'Entities', url: null, onClick: () => this.setArenaDisplay('Entities') },
+                                { name: 'Actions', key: 'Actions', url: null, onClick: () => this.setArenaDisplay('Actions') },
+                                { name: 'Train Dialogs', key: 'TrainDialogs', url: null, onClick: () => this.setArenaDisplay('TrainDialogs') },
+                                { name: 'Log Dialogs', key: 'LogDialogs', url: null, onClick: () => this.setArenaDisplay('LogDialogs') }
+                            ]
+                        }]}
+                    />
+                </div>
+                <div className="tgbackToAppsDiv">
+                    <Link className="backToApps" onClick={() => this.props.setDisplayMode(DisplayMode.AppList)}><span className="ms-Icon ms-Icon--Back backToApps backToAppsIcon" aria-hidden="true"></span>&nbsp;&nbsp;App List</Link>
                 </div>
             </div>
         )
@@ -139,9 +133,9 @@ class AppAdmin extends React.Component<Props, ComponentState> {
     }
     render() {
         return (
-            <div className="container">
-                {this.props.display.displayWebchat == true ?
-                    this.renderWithEmulator()
+            <div>
+                {this.props.display.displayWebchat == true
+                    ? this.renderWithEmulator()
                     : this.renderWithoutEmulator()
                 }
             </div>
